==> 코드 정리/1. v2.ipynb <==
# v2

import pandas as pd
import numpy as np

def ChangeDirection(x):  # 방위각을 기준으로 유향(wd) 바꾸기
    if x <= 11.25:
        return 'N'
    elif (x > 11.25) & (x <= 33.75):
        return 'NNE'
    elif (x > 33.75) & (x <= 56.25):
        return 'NE'
    elif (x > 56.25) & (x <= 78.75):
        return 'ENE'
    elif (x > 78.75) & (x <= 101.25):
        return 'E'
    elif (x > 101.25) & (x <= 123.75):
        return 'ESE'
    elif (x > 123.75) & (x <= 146.25):
        return 'SE'
    elif (x > 146.25) & (x <= 168.75):
        return 'SSE'
    elif (x > 168.75) & (x <= 191.25):
        return 'S'
    elif (x > 191.25) & (x <= 213.75):
        return 'SSW'
    elif (x > 213.75) & (x <= 236.25):
        return 'SW'
    elif (x > 236.25) & (x <= 258.75):
        return 'WSW'
    elif (x > 258.75) & (x <= 281.25):
        return 'W'
    elif (x > 281.25) & (x <= 303.75):
        return 'WNW'
    elif (x > 303.75) & (x <= 326.25):
        return 'NW'
    elif (x > 326.25) & (x <= 348.75):
        return 'NNW'
    elif (x > 348.75) & (x <= 360):
        return 'N'

### ulsan_anch_train_final

### ulsan_anch_train_final #######################################################################################################
ulsan_anch_train_final = pd.read_csv('./원본데이터/ulsan_anch_train_final.csv')


# 시간 전처리
import re
for i in range(len(ulsan_anch_train_final)):
    var1 = ulsan_anch_train_final.loc[i,'ulsan_anch_train_final.time']
    var1 = var1.replace('"','')
    var1 = re.split('-| |:', var1)
    if len(var1) == 6:
        var_year = var1[0]
        var_month = var1[1]
        var_day = var1[2]
        var_hour = var1[3]
        var_min = var1[4]
        var_sec = var1[5]
        ulsan_anch_train_final.loc[i,'ulsan_anch_train_final.year'] = var_year
        ulsan_anch_train_final.loc[i,'ulsan_anch_train_final.month'] = var_month
        ulsan_anch_train_final.loc[i,'ulsan_anch_train_final.day'] = var_day
        ulsan_anch_train_final.loc[i,'ulsan_anch_train_final.hour'] = var_hour
        ulsan_anch_train_final.loc[i,'ulsan_anch_train_final.min'] = var_min
        ulsan_anch_train_final.loc[i,'ulsan_anch_train_final.sec'] = var_sec
    else:
        var_year = var1[0]
        var_month = var1[1]
        var_day = var1[2]
        var_hour = var1[3]
        var_min = var1[4]
        var_sec == '00'
        ulsan_anch_train_final.loc[i,'ulsan_anch_train_final.year'] = var_year
        ulsan_anch_train_final.loc[i,'ulsan_anch_train_final.month'] = var_month
        ulsan_anch_train_final.loc[i,'ulsan_anch_train_final.day'] = var_day
        ulsan_anch_train_final.loc[i,'ulsan_anch_train_final.hour'] = var_hour
        ulsan_anch_train_final.loc[i,'ulsan_anch_train_final.min'] = var_min
        ulsan_anch_train_final.loc[i,'ulsan_anch_train_final.sec'] = var_sec

# 위도 전처리 - 숫자만 출력
for i in range(len(ulsan_anch_train_final)):
    var1 = ulsan_anch_train_final.loc[i,'ulsan_anch_train_final.latitude']
    var1 = var1.replace('N','')
    var1 = var1.replace('"','')
    ulsan_anch_train_final.loc[i,'ulsan_anch_train_final.lat'] = var1

# 경도 전처리 - 숫자만 출력
for i in range(len(ulsan_anch_train_final)):
    var1 = ulsan_anch_train_final.loc[i,'ulsan_anch_train_final.longitude']
    var1 = var1.replace('E','')
    var1 = var1.replace('"','')
    ulsan_anch_train_final.loc[i,'ulsan_anch_train_final.lon'] = var1

# 불필요한 컬럼 삭제
ulsan_anch_train_final = ulsan_anch_train_final.drop(ulsan_anch_train_final.columns[[0]],axis=1)
ulsan_anch_train_final = ulsan_anch_train_final.drop(columns = ['ulsan_anch_train_final.time', 'ulsan_anch_train_final.latitude', 'ulsan_anch_train_final.longitude'])

# .cog 컬럼의 결측치(-99.9)를 평균값으로 바꾸기
var_mean4 = ulsan_anch_train_final['ulsan_anch_train_final.cog'].mean()
var_count4 = (ulsan_anch_train_final['ulsan_anch_train_final.cog'] == -99.9)
ulsan_anch_train_final.loc[var_count4,'ulsan_anch_train_final.cog'] = var_mean4

# csv 파일로 저장
ulsan_anch_train_final.to_csv("./v2/ulsan_anch_train_final_v2.csv", index = False)

### ulsan_drag_train_final

### ulsan_drag_train_final #######################################################################################################
ulsan_drag_train_final = pd.read_csv('./원본데이터/ulsan_drag_train_final.csv')

# 시간 전처리
import re
for i in range(len(ulsan_drag_train_final)):
    var1 = ulsan_drag_train_final.loc[i,'ulsan_drag_train_final.time']
    var1 = var1.replace('"','')
    var1 = re.split('-| |:', var1)
    if len(var1) == 6:
        var_year = var1[0]
        var_month = var1[1]
        var_day = var1[2]
        var_hour = var1[3]
        var_min = var1[4]
        var_sec = var1[5]
        ulsan_drag_train_final.loc[i,'ulsan_drag_train_final.year'] = var_year
        ulsan_drag_train_final.loc[i,'ulsan_drag_train_final.month'] = var_month
        ulsan_drag_train_final.loc[i,'ulsan_drag_train_final.day'] = var_day
        ulsan_drag_train_final.loc[i,'ulsan_drag_train_final.hour'] = var_hour
        ulsan_drag_train_final.loc[i,'ulsan_drag_train_final.min'] = var_min
        ulsan_drag_train_final.loc[i,'ulsan_drag_train_final.sec'] = var_sec
    else:
        var_year = var1[0]
        var_month = var1[1]
        var_day = var1[2]
        var_hour = var1[3]
        var_min = var1[4]
        var_sec == '00'
        ulsan_drag_train_final.loc[i,'ulsan_drag_train_final.year'] = var_year
        ulsan_drag_train_final.loc[i,'ulsan_drag_train_final.month'] = var_month
        ulsan_drag_train_final.loc[i,'ulsan_drag_train_final.day'] = var_day
        ulsan_drag_train_final.loc[i,'ulsan_drag_train_final.hour'] = var_hour
        ulsan_drag_train_final.loc[i,'ulsan_drag_train_final.min'] = var_min
        ulsan_drag_train_final.loc[i,'ulsan_drag_train_final.sec'] = var_sec

# 위도 전처리 - 숫자만 출력
for i in range(len(ulsan_drag_train_final)):
    var1 = ulsan_drag_train_final.loc[i,'ulsan_drag_train_final.latitude']
    var1 = var1.replace('N','')
    var1 = var1.replace('"','')
    ulsan_drag_train_final.loc[i,'ulsan_drag_train_final.lat'] = var1

# 경도 전처리 - 숫자만 출력
for i in range(len(ulsan_drag_train_final)):
    var1 = ulsan_drag_train_final.loc[i,'ulsan_drag_train_final.longitude']
    var1 = var1.replace('E','')
    var1 = var1.replace('"','')
    ulsan_drag_train_final.loc[i,'ulsan_drag_train_final.lon'] = var1

# 불필요한 컬럼 삭제
ulsan_drag_train_final = ulsan_drag_train_final.drop(ulsan_drag_train_final.columns[[0]],axis=1)
ulsan_drag_train_final = ulsan_drag_train_final.drop(columns = ['ulsan_drag_train_final.time', 'ulsan_drag_train_final.latitude', 'ulsan_drag_train_final.longitude'])

# .cog 컬럼의 결측치(-99.9)를 평균값으로 바꾸기
var_mean2 = ulsan_drag_train_final['ulsan_drag_train_final.cog'].mean()
var_count2 = (ulsan_drag_train_final['ulsan_drag_train_final.cog'] == -99.9)
ulsan_drag_train_final.loc[var_count2,'ulsan_drag_train_final.cog'] = var_mean2

# .hdg 컬럼의 결측치(-99)를 평균값으로 바꾸기
var_mean3 = ulsan_drag_train_final['ulsan_drag_train_final.hdg'].mean()
var_count3 = (ulsan_drag_train_final['ulsan_drag_train_final.hdg'] == -99)
ulsan_drag_train_final.loc[var_count3,'ulsan_drag_train_final.hdg'] = var_mean3

# csv 파일로 저장
ulsan_drag_train_final.to_csv("./v2/ulsan_drag_train_final_v2.csv", index = False)

### ulsan_anch_drag_test

### ulsan_anch_drag_test #######################################################################################################
ulsan_anch_drag_test = pd.read_csv('./원본데이터/ulsan_anch_drag_test.csv')

## 시간 전처리
import re
for i in range(len(ulsan_anch_drag_test)):
    var1 = ulsan_anch_drag_test.loc[i,'ulsan_anch_drag_test.time']
    var1 = var1.replace('"','')
    var1 = re.split('-| |:', var1)
    if len(var1) == 6:
        var_year = var1[0]
        var_month = var1[1]
        var_day = var1[2]
        var_hour = var1[3]
        var_min = var1[4]
        var_sec = var1[5]
        ulsan_anch_drag_test.loc[i,'year'] = var_year
        ulsan_anch_drag_test.loc[i,'month'] = var_month
        ulsan_anch_drag_test.loc[i,'day'] = var_day
        ulsan_anch_drag_test.loc[i,'hour'] = var_hour
        ulsan_anch_drag_test.loc[i,'min'] = var_min
        ulsan_anch_drag_test.loc[i,'sec'] = var_sec
    else:
        var_year = var1[0]
        var_month = var1[1]
        var_day = var1[2]
        var_hour = var1[3]
        var_min = var1[4]
        var_sec == '00'
        ulsan_anch_drag_test.loc[i,'year'] = var_year
        ulsan_anch_drag_test.loc[i,'month'] = var_month
        ulsan_anch_drag_test.loc[i,'day'] = var_day
        ulsan_anch_drag_test.loc[i,'hour'] = var_hour
        ulsan_anch_drag_test.loc[i,'min'] = var_min
        ulsan_anch_drag_test.loc[i,'sec'] = var_sec

# 위도 전처리 - 숫자만 출력
for i in range(len(ulsan_anch_drag_test)):
    var1 = ulsan_anch_drag_test.loc[i,'ulsan_anch_drag_test.lat']
    var1 = var1.replace('N','')
    var1 = var1.replace('"','')
    ulsan_anch_drag_test.loc[i,'ulsan_anch_drag_test.lat'] = var1

# 경도 전처리 - 숫자만 출력
for i in range(len(ulsan_anch_drag_test)):
    var1 = ulsan_anch_drag_test.loc[i,'ulsan_anch_drag_test.lon']
    var1 = var1.replace('E','')
    var1 = var1.replace('"','')
    ulsan_anch_drag_test.loc[i,'ulsan_anch_drag_test.lon'] = var1

# 불필요한 컬럼 삭제
ulsan_anch_drag_test = ulsan_anch_drag_test.drop(ulsan_anch_drag_test.columns[[0]],axis=1)
ulsan_anch_drag_test = ulsan_anch_drag_test.drop(columns = ['ulsan_anch_drag_test.time'])

# .cog 컬럼의 결측치(-99.9)를 평균값으로 바꾸기
var_mean = ulsan_anch_drag_test['ulsan_anch_drag_test.cog'].mean()
var_count = (ulsan_anch_drag_test['ulsan_anch_drag_test.cog'] == -99.9)
ulsan_anch_drag_test.loc[var_count,'ulsan_anch_drag_test.cog'] = var_mean

# csv 파일로 저장
ulsan_anch_drag_test.to_csv("./v2/ulsan_anch_drag_test_v2.csv", index = False)

### khnp_buoy

### khnp_buoy_train #######################################################################################################
KHNP_Buoy_train = pd.read_csv("./원본데이터/KHNP_Buoy_train.csv")

## 불필요한 컬럼 삭제. (Unnamed 컬럼 삭제)
KHNP_Buoy_train = KHNP_Buoy_train.drop(KHNP_Buoy_train.columns[[0]],axis=1)

## 시간 컬럼 분할 (연/월/일/시/분)
KHNP_Buoy_train['khnp_buoy_train.yyyymmddhhmi'] = KHNP_Buoy_train['khnp_buoy_train.yyyymmddhhmi'].astype('string')

KHNP_Buoy_train['khnp_buoy_train.year'] = KHNP_Buoy_train['khnp_buoy_train.yyyymmddhhmi'].str.slice(0,4)  # 년
KHNP_Buoy_train['khnp_buoy_train.month'] = KHNP_Buoy_train['khnp_buoy_train.yyyymmddhhmi'].str.slice(4,6) # 월
KHNP_Buoy_train['khnp_buoy_train.day'] = KHNP_Buoy_train['khnp_buoy_train.yyyymmddhhmi'].str.slice(6,8)   # 일
KHNP_Buoy_train['khnp_buoy_train.hour'] = KHNP_Buoy_train['khnp_buoy_train.yyyymmddhhmi'].str.slice(8,10) # 시
KHNP_Buoy_train['khnp_buoy_train.min'] = KHNP_Buoy_train['khnp_buoy_train.yyyymmddhhmi'].str.slice(10,12) # 분

## 시간 컬럼 삭제
KHNP_Buoy_train = KHNP_Buoy_train.drop(columns=['khnp_buoy_train.yyyymmddhhmi'],axis=1)

## train .ws 컬럼 결측치(-999) 처리 : nan 값으로 변경 후 평균값으로 대체
KHNP_Buoy_train[KHNP_Buoy_train['khnp_buoy_train.ws']==-999]
KHNP_Buoy_train.loc[KHNP_Buoy_train['khnp_buoy_train.ws'] == -999, 'khnp_buoy_train.ws'] = np.nan  # 결측치(-999)를 NaN으로 바꾸기
KHNP_Buoy_train.fillna(round(KHNP_Buoy_train['khnp_buoy_train.ws'].mean(numeric_only=True), 1), inplace = True) # NaN을 평균 값으로 바꾸기

## train .wd 컬럼 결측치(-999) 처리 : nan 값으로 변경 후 16방위로 대체
KHNP_Buoy_train[KHNP_Buoy_train['khnp_buoy_train.wd']==-999]
KHNP_Buoy_train.loc[KHNP_Buoy_train['khnp_buoy_train.wd'] == -999, 'khnp_buoy_train.wd'] = np.nan  # 결측치(-999)를 NaN으로 바꾸기
KHNP_Buoy_train['khnp_buoy_train.wd'] = KHNP_Buoy_train['khnp_buoy_train.wd'].map(ChangeDirection)
KHNP_Buoy_train.fillna("X", inplace = True)  # NaN을 'X'로 바꾸기 (공백으로 두었을 때, csv로 변환하면 그대로 null값이되어 'X'로 설정)    

## 1시간별 유향 최빈값
# 년, 월, 일, 시를 기준으로 최빈값 구하기 
khnp_buoy_train_wd_mode = KHNP_Buoy_train.loc[:,'khnp_buoy_train.ws':].groupby(['khnp_buoy_train.year', 'khnp_buoy_train.month', 'khnp_buoy_train.day', 'khnp_buoy_train.hour'])['khnp_buoy_train.wd'].agg(lambda x: pd.Series.mode(x)[0]).to_frame()
khnp_buoy_train_wd_mode = khnp_buoy_train_wd_mode.T.unstack().reset_index()
khnp_buoy_train_wd_mode.drop(columns = ['level_4'], inplace = True)
khnp_buoy_train_wd_mode = khnp_buoy_train_wd_mode.rename(columns = {0:'khnp_buoy_train.wd.mode'})

## KHNP_Buoy_train 데이터프레임에 khnp_buoy_train.wd.mode 컬럼 추가
df_train = pd.merge(KHNP_Buoy_train, khnp_buoy_train_wd_mode, how='left')
df_train.loc[df_train['khnp_buoy_train.wd']=="X", 'khnp_buoy_train.wd'] = df_train['khnp_buoy_train.wd.mode'] # 유향의 결측치를 시 기준 최빈값으로 채우기, 결측치가 최빈값이면 그대로 'X' 값이 채워짐.

khnp_buoy_train_v2 = df_train
khnp_buoy_train_v2 = khnp_buoy_train_v2.drop(columns = ['khnp_buoy_train.wd.mode'])

## csv 파일로 저장
khnp_buoy_train_v2.to_csv("./v2/khnp_buoy_train_v2.csv", index=False)


### khnp_buoy_test #######################################################################################################
KHNP_Buoy_test =pd.read_csv("./원본데이터/KHNP_Buoy_test.csv")

## 불필요한 컬럼 삭제 (Unnamed 컬럼 삭제)
KHNP_Buoy_test = KHNP_Buoy_test.drop(KHNP_Buoy_test.columns[[0]],axis=1)

## 시간 컬럼 분할 (연/월/일/시/분)
KHNP_Buoy_test['khnp_buoy_test.yyyymmddhhmi'] = KHNP_Buoy_test['khnp_buoy_test.yyyymmddhhmi'].astype('string')

KHNP_Buoy_test['khnp_buoy_test.year'] = KHNP_Buoy_test['khnp_buoy_test.yyyymmddhhmi'].str.slice(0,4)  # 년
KHNP_Buoy_test['khnp_buoy_test.month'] = KHNP_Buoy_test['khnp_buoy_test.yyyymmddhhmi'].str.slice(4,6) # 월
KHNP_Buoy_test['khnp_buoy_test.day'] = KHNP_Buoy_test['khnp_buoy_test.yyyymmddhhmi'].str.slice(6,8)   # 일
KHNP_Buoy_test['khnp_buoy_test.hour'] = KHNP_Buoy_test['khnp_buoy_test.yyyymmddhhmi'].str.slice(8,10) # 시
KHNP_Buoy_test['khnp_buoy_test.min'] = KHNP_Buoy_test['khnp_buoy_test.yyyymmddhhmi'].str.slice(10,12) # 분

## 시간 컬럼 삭제
KHNP_Buoy_test = KHNP_Buoy_test.drop(columns=['khnp_buoy_test.yyyymmddhhmi'],axis=1)

## test .ws컬럼 결측치(-999) 처리 : nan 값으로 변경 후 평균값으로 대체
KHNP_Buoy_test[KHNP_Buoy_test['khnp_buoy_test.ws']==-999]
KHNP_Buoy_test.loc[KHNP_Buoy_test['khnp_buoy_test.ws'] == -999, 'khnp_buoy_test.ws'] = np.nan  # 결측치(-999)를 NaN으로 바꾸기
KHNP_Buoy_test.fillna(round(KHNP_Buoy_test['khnp_buoy_test.ws'].mean(numeric_only=True),1), inplace = True) # NaN을 평균 값으로 바꾸기

## train .wd컬럼 결측치(-999) 처리 : nan 값으로 변경 후 16방위로 대체
KHNP_Buoy_test[KHNP_Buoy_test['khnp_buoy_test.wd']==-999]
KHNP_Buoy_test.loc[KHNP_Buoy_test['khnp_buoy_test.wd'] == -999, 'khnp_buoy_test.wd'] = np.nan # 결측치(-999)를 NaN으로 바꾸기
KHNP_Buoy_test['khnp_buoy_test.wd'] = KHNP_Buoy_test['khnp_buoy_test.wd'].map(ChangeDirection)
KHNP_Buoy_test.fillna("X", inplace = True) # NaN을 'X'로 바꾸기 (공백으로 두었을 때, csv로 변환하면 그대로 null값이되어 'X'로 설정)

## 1시간별 유향 최빈값
# 년, 월, 일, 시를 기준으로 최빈값 구하기 
khnp_buoy_test_wd_mode = KHNP_Buoy_test.loc[:,'khnp_buoy_test.ws':].groupby(['khnp_buoy_test.year', 'khnp_buoy_test.month', 'khnp_buoy_test.day', 'khnp_buoy_test.hour'])['khnp_buoy_test.wd'].agg(lambda x: pd.Series.mode(x)[0]).to_frame()
khnp_buoy_test_wd_mode = khnp_buoy_test_wd_mode.T.unstack().reset_index()
khnp_buoy_test_wd_mode.drop(columns = ['level_4'], inplace = True)
khnp_buoy_test_wd_mode = khnp_buoy_test_wd_mode.rename(columns = {0:'khnp_buoy_test.wd.mode'})

## KHNP_Buoy_test 데이터프레임에 khnp_buoy_test.wd.mode 컬럼 추가
df_test = pd.merge(KHNP_Buoy_test, khnp_buoy_test_wd_mode, how='left')
df_test.loc[df_test['khnp_buoy_test.wd']=="X", 'khnp_buoy_test.wd'] = df_test['khnp_buoy_test.wd.mode']

khnp_buoy_test_v2 = df_test
khnp_buoy_test_v2 = khnp_buoy_test_v2.drop(columns = ['khnp_buoy_test.wd.mode'])

## csv 파일로 저장
khnp_buoy_test_v2.to_csv("./v2/khnp_buoy_test_v2.csv", index= False)

### kma_lightbecon

### kma_lightbecon_train #######################################################################################################
df_train = pd.read_csv('./원본데이터/kma_lightbecon_train.csv')

# 불필요한 컬럼 삭제
df_train.drop(labels='kma_lightbecon_train.stn', axis=1, inplace=True)
df_train.drop(['Unnamed: 0'], axis=1, inplace=True)

# 시간 컬럼 분할 (연/월/일/시/분)
df_train['kma_lightbecon_train.yyyymmddhhmi'] = df_train['kma_lightbecon_train.yyyymmddhhmi'].astype('string')

df_train['kma_lightbecon_train.year'] = df_train['kma_lightbecon_train.yyyymmddhhmi'].str.slice(0,4)  # 년
df_train['kma_lightbecon_train.month'] = df_train['kma_lightbecon_train.yyyymmddhhmi'].str.slice(4,6) # 월
df_train['kma_lightbecon_train.day'] = df_train['kma_lightbecon_train.yyyymmddhhmi'].str.slice(6,8)   # 일
df_train['kma_lightbecon_train.hour'] = df_train['kma_lightbecon_train.yyyymmddhhmi'].str.slice(8,10) # 시
df_train['kma_lightbecon_train.min'] = df_train['kma_lightbecon_train.yyyymmddhhmi'].str.slice(10,12) # 분

# 시간 컬럼 삭제
df_train.drop(labels='kma_lightbecon_train.yyyymmddhhmi', axis=1, inplace=True)

# 특정 인덱스명 가진 행 모두 제거 (울산('Ideockseo')만 필터링)
idx = df_train[df_train['kma_lightbecon_train.stn_name']=='Oryuk Island'].index
df_train = df_train.drop(idx)

# 중복값 가진 행 제거
df_train = df_train.drop_duplicates(subset=['kma_lightbecon_train.month','kma_lightbecon_train.day','kma_lightbecon_train.hour','kma_lightbecon_train.min']) 

#날짜(년,월,시,분) 오름차순 정렬
df_train = df_train.sort_values(by=['kma_lightbecon_train.year','kma_lightbecon_train.month','kma_lightbecon_train.day','kma_lightbecon_train.hour','kma_lightbecon_train.min'], ascending=[True, True, True, True, True])

#저장
df_train.to_csv("./v2/kma_lightbecon_train_v2.csv", index=False)



### kma_lightbecon_test  #######################################################################################################
df_test = pd.read_csv('./원본데이터/kma_lightbecon_test.csv')

# 불필요한 컬럼 삭제
df_test.drop(labels='kma_lightbecon_test.stn', axis=1, inplace=True)
df_test.drop(['Unnamed: 0'], axis=1, inplace=True)

# 시간 컬럼 분할 (연/월/일/시/분)
df_test['kma_lightbecon_test.yyyymmddhhmi'] = df_test['kma_lightbecon_test.yyyymmddhhmi'].astype('string')

df_test['kma_lightbecon_test.year'] = df_test['kma_lightbecon_test.yyyymmddhhmi'].str.slice(0,4)     # 년
df_test['kma_lightbecon_test.month'] = df_test['kma_lightbecon_test.yyyymmddhhmi'].str.slice(4,6)    # 월
df_test['kma_lightbecon_test.day'] = df_test['kma_lightbecon_test.yyyymmddhhmi'].str.slice(6,8)      # 일
df_test['kma_lightbecon_test.hour'] = df_test['kma_lightbecon_test.yyyymmddhhmi'].str.slice(8,10)    # 시
df_test['kma_lightbecon_test.min'] = df_test['kma_lightbecon_test.yyyymmddhhmi'].str.slice(10,12) # 분

# 시간 컬럼 삭제
df_test.drop(labels='kma_lightbecon_test.yyyymmddhhmi', axis=1, inplace=True)

# 특정 인덱스명 가진 행 모두 제거 (울산('Ideockseo')만 필터링)
idx = df_test[df_test['kma_lightbecon_test.stn_name']=='Oryuk Island'].index
df_test = df_test.drop(idx)

# 중복값 가진 행 제거
df_test=df_test.drop_duplicates(subset=['kma_lightbecon_test.month','kma_lightbecon_test.day','kma_lightbecon_test.hour','kma_lightbecon_test.min']) 

# 날짜(년,월,시,분) 오름차순 정렬
df_test=df_test.sort_values(by=['kma_lightbecon_test.year','kma_lightbecon_test.month','kma_lightbecon_test.day','kma_lightbecon_test.hour','kma_lightbecon_test.min'], ascending=[True, True, True, True, True])

# csv 파일로 저장
df_test.to_csv("./v2/kma_lightbecon_test_v2.csv", index=False)

### kma_pagobuoy

### kma_pagobuoy_train #######################################################################################################
kma_pagobuoy_train = pd.read_csv('./원본데이터/kma_pagobuoy_train.csv')
df_train = kma_pagobuoy_train

## 울산:간절곶 데이터만 필터링
kma_pagobuoy_train = kma_pagobuoy_train[(kma_pagobuoy_train['kma_pagobuoy_train.stn_name'] == '간절곶') | (kma_pagobuoy_train['kma_pagobuoy_train.stn_name'] == '당사')]

## 불필요/중복 컬럼 (Unnamed: 0, kma_pagobuoy_train/test.stn) 제거
kma_pagobuoy_train = kma_pagobuoy_train.drop(columns=['Unnamed: 0', 'kma_pagobuoy_train.stn'], axis=1)

## 시간 컬럼 분할 (연/월/일/시)
kma_pagobuoy_train['kma_pagobuoy_train.yyyymmddhhmi'] = kma_pagobuoy_train['kma_pagobuoy_train.yyyymmddhhmi'].astype('string')

kma_pagobuoy_train['kma_pagobuoy_train.year'] = kma_pagobuoy_train['kma_pagobuoy_train.yyyymmddhhmi'].str.slice(0,4)  # 년
kma_pagobuoy_train['kma_pagobuoy_train.month'] = kma_pagobuoy_train['kma_pagobuoy_train.yyyymmddhhmi'].str.slice(4,6) # 월
kma_pagobuoy_train['kma_pagobuoy_train.day'] = kma_pagobuoy_train['kma_pagobuoy_train.yyyymmddhhmi'].str.slice(6,8)   # 일
kma_pagobuoy_train['kma_pagobuoy_train.hour'] = kma_pagobuoy_train['kma_pagobuoy_train.yyyymmddhhmi'].str.slice(8,10) # 시

## 시간 컬럼 제거
kma_pagobuoy_train = kma_pagobuoy_train.drop(columns=['kma_pagobuoy_train.yyyymmddhhmi'], axis=1)

## train 결측치(-99.9) 처리 : nan 값으로 변경 후 평균값으로 대체
# .max_wh 컬럼 결측치(-99.9) 처리
kma_pagobuoy_train[kma_pagobuoy_train['kma_pagobuoy_train.max_wh']==-99.9]
kma_pagobuoy_train.loc[kma_pagobuoy_train['kma_pagobuoy_train.max_wh'] == -99.9, 'kma_pagobuoy_train.max_wh'] = np.nan
kma_pagobuoy_train.fillna(round(df_train.mean(numeric_only=True), 1), inplace = True)

# .sig_wh 컬럼 결측치(-99.9) 처리
kma_pagobuoy_train[kma_pagobuoy_train['kma_pagobuoy_train.sig_wh']==-99.9]
kma_pagobuoy_train.loc[kma_pagobuoy_train['kma_pagobuoy_train.sig_wh'] == -99.9, 'kma_pagobuoy_train.sig_wh'] = np.nan
kma_pagobuoy_train.fillna(round(df_train.mean(numeric_only=True), 1), inplace = True)

# .mean_wh 컬럼 결측치(-99.9) 처리
kma_pagobuoy_train[kma_pagobuoy_train['kma_pagobuoy_train.mean_wh']==-99.9]
kma_pagobuoy_train.loc[kma_pagobuoy_train['kma_pagobuoy_train.mean_wh'] == -99.9, 'kma_pagobuoy_train.mean_wh'] = np.nan
kma_pagobuoy_train.fillna(round(df_train.mean(numeric_only=True), 1), inplace = True)

## csv 파일로 저장
kma_pagobuoy_train.to_csv('../코드 정리/v2/kma_pagobuoy_train_v2.csv', index=False)

### kma_pagobuoy_test ##########################################################################################################
kma_pagobuoy_test = pd.read_csv('./원본데이터/kma_pagobuoy_test.csv')
df_test = kma_pagobuoy_test

## 울산:간절곶 데이터만 필터링
kma_pagobuoy_test = kma_pagobuoy_test[(kma_pagobuoy_test['kma_pagobuoy_test.stn_name'] == '간절곶') | (kma_pagobuoy_test['kma_pagobuoy_test.stn_name'] == '당사')]

## 불필요/중복 컬럼 (Unnamed: 0, kma_pagobuoy_train/test.stn) 제거
kma_pagobuoy_test = kma_pagobuoy_test.drop(columns=['Unnamed: 0', 'kma_pagobuoy_test.stn'], axis=1)

## 시간 컬럼 분할 (연/월/일/시)
kma_pagobuoy_test['kma_pagobuoy_test.yyyymmddhhmi'] = kma_pagobuoy_test['kma_pagobuoy_test.yyyymmddhhmi'].astype('string')

kma_pagobuoy_test['kma_pagobuoy_test.year'] = kma_pagobuoy_test['kma_pagobuoy_test.yyyymmddhhmi'].str.slice(0,4)  # 년
kma_pagobuoy_test['kma_pagobuoy_test.month'] = kma_pagobuoy_test['kma_pagobuoy_test.yyyymmddhhmi'].str.slice(4,6) # 월
kma_pagobuoy_test['kma_pagobuoy_test.day'] = kma_pagobuoy_test['kma_pagobuoy_test.yyyymmddhhmi'].str.slice(6,8)   # 일
kma_pagobuoy_test['kma_pagobuoy_test.hour'] = kma_pagobuoy_test['kma_pagobuoy_test.yyyymmddhhmi'].str.slice(8,10) # 시

## 시간 컬럼 제거
kma_pagobuoy_test = kma_pagobuoy_test.drop(columns=['kma_pagobuoy_test.yyyymmddhhmi'], axis=1)

## test 결측치(-99.9) 처리 : nan 값으로 변경 후 평균값으로 대체
# .max_wh 컬럼 결측치(-99.9) 처리
kma_pagobuoy_test[kma_pagobuoy_test['kma_pagobuoy_test.max_wh']== -99.9]
kma_pagobuoy_test.loc[kma_pagobuoy_test['kma_pagobuoy_test.max_wh'] == -99.9, 'kma_pagobuoy_test.max_wh'] = np.nan
kma_pagobuoy_test.fillna(round(df_test.mean(numeric_only=True), 1), inplace = True)

# .sig_wh 컬럼 결측치(-99.9) 처리
kma_pagobuoy_test[kma_pagobuoy_test['kma_pagobuoy_test.sig_wh']== -99.9]
kma_pagobuoy_test.loc[kma_pagobuoy_test['kma_pagobuoy_test.sig_wh'] == -99.9, 'kma_pagobuoy_test.sig_wh'] = np.nan
kma_pagobuoy_test.fillna(round(df_test.mean(numeric_only=True), 1), inplace = True)

# .mean_wh 컬럼 결측치(-99.9) 처리
kma_pagobuoy_test[kma_pagobuoy_test['kma_pagobuoy_test.mean_wh']== -99.9]
kma_pagobuoy_test.loc[kma_pagobuoy_test['kma_pagobuoy_test.mean_wh'] == -99.9, 'kma_pagobuoy_test.mean_wh'] = np.nan
kma_pagobuoy_test.fillna(round(df_test.mean(numeric_only=True), 1), inplace = True)

## csv 파일로 저장
kma_pagobuoy_test.to_csv('./v2/kma_pagobuoy_test_v2.csv', index=False)
