==> 코드 정리/3. df_train_v2, df_test_v2.ipynb <==
import pandas as pd
import numpy as np

# df_train_v2

df_train = pd.read_csv('./통합데이터/df_train.csv')

#15방향(유향) 문자 -> 숫자 변환
mapping = {"N": 0, "NNE": 1, "NE": 2, "ENE": 3, "E": 4, "ESE": 5, "SE": 6,
           "SSE": 7, "S": 8, "SSW": 9, "SW": 10, "WSW": 11, "W": 12,
           "WNW": 13, "NW": 14, "NNW": 15, "X": 16}

df_train['khnp_buoy_train.wd'] = df_train['khnp_buoy_train.wd'].map(mapping)
df_train['kma_lightbecon_train.wd'] = df_train['kma_lightbecon_train.wd'].map(mapping)

# 트레인데이터 선박번호 4자리 맞춰서 작업하는 코드
df_train['num'] = df_train['num'].apply(lambda x : str(x).zfill(3))
df_train['num'] = df_train['answer'].astype(str) + df_train['num'].astype(str)

# 불필요한 컬럼 삭제
df_train_drop = df_train.drop(columns = ['class', 'khnp_buoy_train.stn_name', 'kma_lightbecon_train.stn_name'], axis = 1)

# null이 존재하는 행 모두 삭제
df_train_drop = df_train_drop.dropna(axis = 0)

# csv 파일로 저장
df_train_drop.to_csv('./통합데이터/df_train_v2.csv', index = False)

# df_test_v2

df_test = pd.read_csv('./통합데이터/df_test.csv')

# 불필요한 컬럼 제거
df_test.drop(labels=['area','khnp_buoy_test.stn_name','kma_lightbecon_test.stn_name'], axis=1, inplace=True)

# 15방향(유향) 문자 -> 숫자 변환 
mapping = {"N": 0, "NNE": 1, "NE": 2, "ENE": 3, "E": 4, "ESE": 5, "SE": 6,
           "SSE": 7, "S": 8, "SSW": 9, "SW": 10, "WSW": 11, "W": 12,
           "WNW": 13, "NW": 14, "NNW": 15, "X": 16}

df_test['khnp_buoy_test.wd'] = df_test['khnp_buoy_test.wd'].map(mapping)
df_test['kma_lightbecon_test.wd'] = df_test['kma_lightbecon_test.wd'].map(mapping)

#'num'컬럼 천의 자리 '1'제거 
df_test['num'] = df_test['num'].apply(lambda x : str(x)[1:])

# null이 존재하는 행 모두 삭제
df_test_drop = df_test.dropna(axis = 0)

# csv 파일로 저장
df_test_drop.to_csv("./통합데이터/df_test_v2.csv", index=False)
