==> 코드 정리/3. df_train_v2_mean, df_test_v2_mean.ipynb <==
import pandas as pd
import numpy as np

# df_train_v2_mean
- df_train 테이블에 시간 데이터에 대한 null 값을 가지는 컬럼들이 존재함
- null을 컬럼의 평균 값으로 대체

df_train = pd.read_csv('./통합데이터/df_train.csv')

## 선박번호 컬럼 재작업(num)
# train 4자리 맞추고 클래스 구분하기 (정박:0, 닻 끌림:1)
df_train['num'] = df_train['num'].apply(lambda x : str(x).zfill(3))
df_train['num'] = df_train['answer'].astype(str) + df_train['num'].astype(str)


## 중복, 필요없는 컬럼 삭제
df_train_drop = df_train.drop(columns=['class', 'khnp_buoy_train.stn_name', 'kma_lightbecon_train.stn_name'], axis=1)

## 유향, 풍향 인코딩
# 방향 딕셔너리 
direction = {"N": 0, "NNE": 1, "NE": 2, "ENE": 3, "E": 4, "ESE": 5, "SE": 6,
           "SSE": 7, "S": 8, "SSW": 9, "SW": 10, "WSW": 11, "W": 12,
           "WNW": 13, "NW": 14, "NNW": 15, "X":16}

df_train_drop['khnp_buoy_train.wd'] = df_train_drop['khnp_buoy_train.wd'].map(direction)
df_train_drop['kma_lightbecon_train.wd'] = df_train_drop['kma_lightbecon_train.wd'].map(direction)

## 결측치 평균으로 채우기
df_train_drop['kma_lightbecon_train.ws'].fillna(df_train_drop['kma_lightbecon_train.ws'].mean(), inplace=True)
df_train_drop['kma_pagobuoy_train.max_wh'].fillna(df_train_drop['kma_pagobuoy_train.max_wh'].mean(), inplace=True)
df_train_drop['kma_pagobuoy_train.sig_wh'].fillna(df_train_drop['kma_pagobuoy_train.sig_wh'].mean(), inplace=True)
df_train_drop['kma_pagobuoy_train.mean_wh'].fillna(df_train_drop['kma_pagobuoy_train.mean_wh'].mean(), inplace=True)

## 방향 결측치는 16으로 채우기 (인코딩)
df_train_drop['kma_lightbecon_train.wd'].fillna(16, inplace=True)


# 피처간 상관관계 분석 (히트맵)
df_hitmap = df_train_drop.loc[:, ['lat', 'lon', 'sog', 'cog', 'hdg', 'answer',
                      'khnp_buoy_train.ws', 'khnp_buoy_train.wd',
                      'kma_lightbecon_train.ws','kma_lightbecon_train.wd',
                      'kma_pagobuoy_train.max_wh', 'kma_pagobuoy_train.sig_wh', 'kma_pagobuoy_train.mean_wh']]                      
df_hitmap = df_hitmap.corr()
df_hitmap.style.background_gradient(cmap='gray')

# csv 파일로 저장
df_train_drop.to_csv('./통합데이터/df_train_v2_mean.csv', index=False)

# df_test_v2_mean
- df_test 테이블에 시간 데이터에 대한 null 값을 가지는 컬럼들이 존재함
- null을 컬럼의 평균 값으로 대체

df_test = pd.read_csv('./통합데이터/df_test.csv')

## 선박번호 컬럼 재작업(num)
# test 3자리 맞추기
df_test['num'] = df_test['num'].apply(lambda x : str(x)[1:])

## 중복, 필요없는 컬럼 삭제
df_test_drop = df_test.drop(columns=['area', 'khnp_buoy_test.stn_name', 'kma_lightbecon_test.stn_name'], axis=1)

## 유향, 풍향 인코딩
# 방향 딕셔너리 
direction = {"N": 0, "NNE": 1, "NE": 2, "ENE": 3, "E": 4, "ESE": 5, "SE": 6,
           "SSE": 7, "S": 8, "SSW": 9, "SW": 10, "WSW": 11, "W": 12,
           "WNW": 13, "NW": 14, "NNW": 15, "X":16}

df_test_drop['khnp_buoy_test.wd'] = df_test_drop['khnp_buoy_test.wd'].map(direction)
df_test_drop['kma_lightbecon_test.wd'] = df_test_drop['kma_lightbecon_test.wd'].map(direction)

# 결측치 평균으로 채우기
df_test_drop['kma_lightbecon_test.ws'].fillna(df_test_drop['kma_lightbecon_test.ws'].mean(), inplace=True)
df_test_drop['kma_pagobuoy_test.max_wh'].fillna(df_test_drop['kma_pagobuoy_test.max_wh'].mean(), inplace=True)
df_test_drop['kma_pagobuoy_test.sig_wh'].fillna(df_test_drop['kma_pagobuoy_test.sig_wh'].mean(), inplace=True)
df_test_drop['kma_pagobuoy_test.mean_wh'].fillna(df_test_drop['kma_pagobuoy_test.mean_wh'].mean(), inplace=True)

# 방향 결측치는 16으로 채우기 (인코딩)
df_test_drop['kma_lightbecon_test.wd'].fillna(16, inplace=True)

# csv 파일로 저장
df_test_drop.to_csv('./통합데이터/df_test_v2_mean.csv', index=False)
