==> 코드 정리/5. df_train_v3_1m, df_test_v3 (1분 단위).ipynb <==
import pandas as pd
import numpy as np

# df_train_v3_1m

### df_train_v3_mean ###############################################################################################
df_train = pd.read_csv("./통합데이터/df_train_v2.csv")

## df_anch_train (answer = 0)
df_anch_train = df_train[(df_train['answer']==0)] # 정박 

# df_lat
df_lat = round(df_anch_train.groupby(['year', 'month', 'day', 'hour', 'min']).mean(),6)['lat'].to_frame()
df_lat = df_lat.T.unstack().reset_index()
df_lat.drop(columns = ['level_5'], inplace = True)
df_lat = df_lat.rename(columns = {0:'lat'})

# df_lon
df_lon = round(df_anch_train.groupby(['year', 'month', 'day', 'hour', 'min']).mean(),6)['lon'].to_frame()
df_lon = df_lon.T.unstack().reset_index()
df_lon.drop(columns = ['level_5'], inplace = True)
df_lon = df_lon.rename(columns = {0:'lon'})

# df_sog
df_sog	= round(df_anch_train.groupby(['year', 'month', 'day', 'hour', 'min']).mean(),1)['sog'].to_frame()
df_sog = df_sog.T.unstack().reset_index()
df_sog.drop(columns = ['level_5'], inplace = True)
df_sog = df_sog.rename(columns = {0:'sog'})

# df_cog
df_cog= round(df_anch_train.groupby(['year', 'month', 'day', 'hour', 'min']).mean(),1)['cog'].to_frame()
df_cog = df_cog.T.unstack().reset_index()
df_cog.drop(columns = ['level_5'], inplace = True)
df_cog = df_cog.rename(columns = {0:'cog'})

# df_hdg
df_hdg = round(df_anch_train.groupby(['year', 'month', 'day', 'hour', 'min']).mean(),1)['hdg'].to_frame()
df_hdg = df_hdg.T.unstack().reset_index()
df_hdg.drop(columns = ['level_5'], inplace = True)
df_hdg = df_hdg.rename(columns = {0:'hdg'})

# df_khnp_buoy_train_ws
df_khnp_buoy_train_ws = round(df_anch_train.groupby(['year', 'month', 'day', 'hour', 'min']).mean(),2)['khnp_buoy_train.ws'].to_frame()
df_khnp_buoy_train_ws = df_khnp_buoy_train_ws.T.unstack().reset_index()
df_khnp_buoy_train_ws.drop(columns = ['level_5'], inplace = True)
df_khnp_buoy_train_ws = df_khnp_buoy_train_ws.rename(columns = {0:'khnp_buoy.ws'})

# df_khnp_buoy_train_wd
df_khnp_buoy_train_wd = df_anch_train.groupby(['year', 'month', 'day', 'hour', 'min'])['khnp_buoy_train.wd'].agg(lambda x: pd.Series.mode(x)[0]).to_frame()
df_khnp_buoy_train_wd = df_khnp_buoy_train_wd.T.unstack().reset_index()
df_khnp_buoy_train_wd.drop(columns = ['level_5'], inplace = True)
df_khnp_buoy_train_wd = df_khnp_buoy_train_wd.rename(columns = {0:'khnp_buoy.wd'})

# df_kma_lightbecon_train_wd
df_kma_lightbecon_train_wd = df_anch_train.groupby(['year', 'month', 'day', 'hour', 'min'])['kma_lightbecon_train.wd'].agg(lambda x: pd.Series.mode(x)[0]).to_frame()
df_kma_lightbecon_train_wd = df_kma_lightbecon_train_wd.T.unstack().reset_index()
df_kma_lightbecon_train_wd.drop(columns = ['level_5'], inplace = True)
df_kma_lightbecon_train_wd = df_kma_lightbecon_train_wd.rename(columns = {0:'kma_lightbecon.wd'})

# df_kma_lightbecon_train_ws
df_kma_lightbecon_train_ws = round(df_anch_train.groupby(['year', 'month', 'day', 'hour', 'min']).mean(),2)['kma_lightbecon_train.ws'].to_frame()
df_kma_lightbecon_train_ws = df_kma_lightbecon_train_ws.T.unstack().reset_index()
df_kma_lightbecon_train_ws.drop(columns = ['level_5'], inplace = True)
df_kma_lightbecon_train_ws = df_kma_lightbecon_train_ws.rename(columns = {0:'kma_lightbecon.ws'})

# df_kma_pagobuoy_train_max_wh
df_kma_pagobuoy_train_max_wh = round(df_anch_train.groupby(['year', 'month', 'day', 'hour', 'min']).mean(),2)['kma_pagobuoy_train.max_wh'].to_frame()
df_kma_pagobuoy_train_max_wh = df_kma_pagobuoy_train_max_wh.T.unstack().reset_index()
df_kma_pagobuoy_train_max_wh.drop(columns = ['level_5'], inplace = True)
df_kma_pagobuoy_train_max_wh = df_kma_pagobuoy_train_max_wh.rename(columns = {0:'kma_pagobuoy.max_wh'})

# df_kma_pagobuoy_train_sig_wh
df_kma_pagobuoy_train_sig_wh = round(df_anch_train.groupby(['year', 'month', 'day', 'hour', 'min']).mean(),2)['kma_pagobuoy_train.sig_wh'].to_frame()
df_kma_pagobuoy_train_sig_wh = df_kma_pagobuoy_train_sig_wh.T.unstack().reset_index()
df_kma_pagobuoy_train_sig_wh.drop(columns = ['level_5'], inplace = True)
df_kma_pagobuoy_train_sig_wh = df_kma_pagobuoy_train_sig_wh.rename(columns = {0:'kma_pagobuoy.sig_wh'})

# df_kma_pagobuoy_train_mean_wh
df_kma_pagobuoy_train_mean_wh = round(df_anch_train.groupby(['year', 'month', 'day', 'hour', 'min']).mean(),2)['kma_pagobuoy_train.mean_wh'].to_frame()
df_kma_pagobuoy_train_mean_wh = df_kma_pagobuoy_train_mean_wh.T.unstack().reset_index()
df_kma_pagobuoy_train_mean_wh.drop(columns = ['level_5'], inplace = True)
df_kma_pagobuoy_train_mean_wh = df_kma_pagobuoy_train_mean_wh.rename(columns = {0:'kma_pagobuoy.mean_wh'})

## Merge
'''
df_lat
df_lon
df_sog
df_cog
df_hdg
df_khnp_buoy_train_ws
df_khnp_buoy_train_wd
df_kma_lightbecon_train_ws
df_kma_lightbecon_train_wd
df_kma_pagobuoy_train_max_wh
df_kma_pagobuoy_train_sig_wh
df_kma_pagobuoy_train_mean_wh
'''
df_0 = pd.merge(df_lat, df_lon, how='inner')
df_0 = pd.merge(df_0, df_sog, how='inner')
df_0 = pd.merge(df_0, df_sog, how='inner')
df_0 = pd.merge(df_0, df_cog, how='inner')
df_0 = pd.merge(df_0, df_hdg, how='inner')
df_0 = pd.merge(df_0, df_khnp_buoy_train_ws, how='inner')
df_0 = pd.merge(df_0, df_khnp_buoy_train_wd, how='inner')
df_0 = pd.merge(df_0, df_kma_lightbecon_train_ws, how='inner')
df_0 = pd.merge(df_0, df_kma_lightbecon_train_wd, how='inner')
df_0 = pd.merge(df_0, df_kma_pagobuoy_train_max_wh, how='inner')
df_0 = pd.merge(df_0, df_kma_pagobuoy_train_sig_wh, how='inner')
df_0 = pd.merge(df_0, df_kma_pagobuoy_train_mean_wh, how='inner')

# merge 완료한 테이블에 answer 컬럼 추가
df_0['answer'] = 0

### df_drag_train (answer = 1) ###############################################################################################
df_drag_train = df_train[(df_train['answer']==1)] # 닻끌림

# df_lat
df_lat = round(df_drag_train.groupby(['year', 'month', 'day', 'hour', 'min']).mean(),6)['lat'].to_frame()
df_lat = df_lat.T.unstack().reset_index()
df_lat.drop(columns = ['level_5'], inplace = True)
df_lat = df_lat.rename(columns = {0:'lat'})

# df_lon
df_lon = round(df_drag_train.groupby(['year', 'month', 'day', 'hour', 'min']).mean(),6)['lon'].to_frame()
df_lon = df_lon.T.unstack().reset_index()
df_lon.drop(columns = ['level_5'], inplace = True)
df_lon = df_lon.rename(columns = {0:'lon'})

### df_sog
df_sog	= round(df_drag_train.groupby(['year', 'month', 'day', 'hour', 'min']).mean(),1)['sog'].to_frame()
df_sog = df_sog.T.unstack().reset_index()
df_sog.drop(columns = ['level_5'], inplace = True)
df_sog = df_sog.rename(columns = {0:'sog'})

### df_cog
df_cog= round(df_drag_train.groupby(['year', 'month', 'day', 'hour', 'min']).mean(),1)['cog'].to_frame()
df_cog = df_cog.T.unstack().reset_index()
df_cog.drop(columns = ['level_5'], inplace = True)
df_cog = df_cog.rename(columns = {0:'cog'})

### df_hdg
df_hdg = round(df_drag_train.groupby(['year', 'month', 'day', 'hour', 'min']).mean(),1)['hdg'].to_frame()
df_hdg = df_hdg.T.unstack().reset_index()
df_hdg.drop(columns = ['level_5'], inplace = True)
df_hdg = df_hdg.rename(columns = {0:'hdg'})

### df_khnp_buoy_train_ws
df_khnp_buoy_train_ws = round(df_drag_train.groupby(['year', 'month', 'day', 'hour', 'min']).mean(),2)['khnp_buoy_train.ws'].to_frame()
df_khnp_buoy_train_ws = df_khnp_buoy_train_ws.T.unstack().reset_index()
df_khnp_buoy_train_ws.drop(columns = ['level_5'], inplace = True)
df_khnp_buoy_train_ws = df_khnp_buoy_train_ws.rename(columns = {0:'khnp_buoy.ws'})

### df_khnp_buoy_train_wd
df_khnp_buoy_train_wd = df_drag_train.groupby(['year', 'month', 'day', 'hour', 'min'])['khnp_buoy_train.wd'].agg(lambda x: pd.Series.mode(x)[0]).to_frame()
df_khnp_buoy_train_wd = df_khnp_buoy_train_wd.T.unstack().reset_index()
df_khnp_buoy_train_wd.drop(columns = ['level_5'], inplace = True)
df_khnp_buoy_train_wd = df_khnp_buoy_train_wd.rename(columns = {0:'khnp_buoy.wd'})

### df_kma_lightbecon_train_wd
df_kma_lightbecon_train_wd = df_drag_train.groupby(['year', 'month', 'day', 'hour', 'min'])['kma_lightbecon_train.wd'].agg(lambda x: pd.Series.mode(x)[0]).to_frame()
df_kma_lightbecon_train_wd = df_kma_lightbecon_train_wd.T.unstack().reset_index()
df_kma_lightbecon_train_wd.drop(columns = ['level_5'], inplace = True)
df_kma_lightbecon_train_wd = df_kma_lightbecon_train_wd.rename(columns = {0:'kma_lightbecon.wd'})

### df_kma_lightbecon_train_ws
df_kma_lightbecon_train_ws = round(df_drag_train.groupby(['year', 'month', 'day', 'hour', 'min']).mean(),2)['kma_lightbecon_train.ws'].to_frame()
df_kma_lightbecon_train_ws = df_kma_lightbecon_train_ws.T.unstack().reset_index()
df_kma_lightbecon_train_ws.drop(columns = ['level_5'], inplace = True)
df_kma_lightbecon_train_ws = df_kma_lightbecon_train_ws.rename(columns = {0:'kma_lightbecon.ws'})

### df_kma_pagobuoy_train_max_wh
df_kma_pagobuoy_train_max_wh = round(df_drag_train.groupby(['year', 'month', 'day', 'hour', 'min']).mean(),2)['kma_pagobuoy_train.max_wh'].to_frame()
df_kma_pagobuoy_train_max_wh = df_kma_pagobuoy_train_max_wh.T.unstack().reset_index()
df_kma_pagobuoy_train_max_wh.drop(columns = ['level_5'], inplace = True)
df_kma_pagobuoy_train_max_wh = df_kma_pagobuoy_train_max_wh.rename(columns = {0:'kma_pagobuoy.max_wh'})

### df_kma_pagobuoy_train_sig_wh
df_kma_pagobuoy_train_sig_wh = round(df_drag_train.groupby(['year', 'month', 'day', 'hour', 'min']).mean(),2)['kma_pagobuoy_train.sig_wh'].to_frame()
df_kma_pagobuoy_train_sig_wh = df_kma_pagobuoy_train_sig_wh.T.unstack().reset_index()
df_kma_pagobuoy_train_sig_wh.drop(columns = ['level_5'], inplace = True)
df_kma_pagobuoy_train_sig_wh = df_kma_pagobuoy_train_sig_wh.rename(columns = {0:'kma_pagobuoy.sig_wh'})

### df_kma_pagobuoy_train_mean_wh
df_kma_pagobuoy_train_mean_wh = round(df_drag_train.groupby(['year', 'month', 'day', 'hour', 'min']).mean(),2)['kma_pagobuoy_train.mean_wh'].to_frame()
df_kma_pagobuoy_train_mean_wh = df_kma_pagobuoy_train_mean_wh.T.unstack().reset_index()
df_kma_pagobuoy_train_mean_wh.drop(columns = ['level_5'], inplace = True)
df_kma_pagobuoy_train_mean_wh = df_kma_pagobuoy_train_mean_wh.rename(columns = {0:'kma_pagobuoy.mean_wh'})

## Merge
'''
df_lat
df_lon
df_sog
df_cog
df_hdg
df_khnp_buoy_train_ws
df_khnp_buoy_train_wd
df_kma_lightbecon_train_ws
df_kma_lightbecon_train_wd
df_kma_pagobuoy_train_max_wh
df_kma_pagobuoy_train_sig_wh
df_kma_pagobuoy_train_mean_wh
'''
df_1 = pd.merge(df_lat, df_lon, how='inner')
df_1 = pd.merge(df_1, df_sog, how='inner')
df_1 = pd.merge(df_1, df_sog, how='inner')
df_1 = pd.merge(df_1, df_cog, how='inner')
df_1 = pd.merge(df_1, df_hdg, how='inner')
df_1 = pd.merge(df_1, df_khnp_buoy_train_ws, how='inner')
df_1 = pd.merge(df_1, df_khnp_buoy_train_wd, how='inner')
df_1 = pd.merge(df_1, df_kma_lightbecon_train_ws, how='inner')
df_1 = pd.merge(df_1, df_kma_lightbecon_train_wd, how='inner')
df_1 = pd.merge(df_1, df_kma_pagobuoy_train_max_wh, how='inner')
df_1 = pd.merge(df_1, df_kma_pagobuoy_train_sig_wh, how='inner')
df_1 = pd.merge(df_1, df_kma_pagobuoy_train_mean_wh, how='inner')

# merge 완료한 테이블에 answer 컬럼 추가
df_1['answer'] = 1

# Concat
df = pd.concat([df_0,df_1], join='inner')

# csv 저장
df.to_csv("./통합데이터/df_train_v3_1m.csv", index=False)

# df

# df_test_v3

# ### df_test_v3_mean 
# df_test = pd.read_csv("./통합데이터/df_test_v2.csv")

# ## df_anch_test
# df_anch_drag_test = df_test # 정박

# ### df_lat
# df_lat = round(df_anch_drag_test.groupby(['year', 'month', 'day', 'hour', 'min']).mean(),6)['lat'].to_frame()
# df_lat = df_lat.T.unstack().reset_index()
# df_lat.drop(columns = ['level_5'], inplace = True)
# df_lat = df_lat.rename(columns = {0:'lat'})

# ### df_lon
# df_lon = round(df_anch_drag_test.groupby(['year', 'month', 'day', 'hour', 'min']).mean(),6)['lon'].to_frame()
# df_lon = df_lon.T.unstack().reset_index()
# df_lon.drop(columns = ['level_5'], inplace = True)
# df_lon = df_lon.rename(columns = {0:'lon'})

# ### df_sog
# df_sog = round(df_anch_drag_test.groupby(['year', 'month', 'day', 'hour', 'min']).mean(),1)['sog'].to_frame()
# df_sog = df_sog.T.unstack().reset_index()
# df_sog.drop(columns = ['level_5'], inplace = True)
# df_sog = df_sog.rename(columns = {0:'sog'})

# ### df_cog
# df_cog= round(df_anch_drag_test.groupby(['year', 'month', 'day', 'hour', 'min']).mean(),1)['cog'].to_frame()
# df_cog = df_cog.T.unstack().reset_index()
# df_cog.drop(columns = ['level_5'], inplace = True)
# df_cog = df_cog.rename(columns = {0:'cog'})

# ### df_hdg
# df_hdg = round(df_anch_drag_test.groupby(['year', 'month', 'day', 'hour', 'min']).mean(),1)['hdg'].to_frame()
# df_hdg = df_hdg.T.unstack().reset_index()
# df_hdg.drop(columns = ['level_5'], inplace = True)
# df_hdg = df_hdg.rename(columns = {0:'hdg'})

# ### df_khnp_buoy_test_ws
# df_khnp_buoy_test_ws = round(df_anch_drag_test.groupby(['year', 'month', 'day', 'hour', 'min']).mean(),2)['khnp_buoy_test.ws'].to_frame()
# df_khnp_buoy_test_ws = df_khnp_buoy_test_ws.T.unstack().reset_index()
# df_khnp_buoy_test_ws.drop(columns = ['level_5'], inplace = True)
# df_khnp_buoy_test_ws = df_khnp_buoy_test_ws.rename(columns = {0:'khnp_buoy.ws'})

# ### df_khnp_buoy_test_wd
# df_khnp_buoy_test_wd = df_anch_drag_test.groupby(['year', 'month', 'day', 'hour','min'])['khnp_buoy_test.wd'].agg(lambda x: pd.Series.mode(x)[0]).to_frame()
# df_khnp_buoy_test_wd = df_khnp_buoy_test_wd.T.unstack().reset_index()
# df_khnp_buoy_test_wd.drop(columns = ['level_5'], inplace = True)
# df_khnp_buoy_test_wd = df_khnp_buoy_test_wd.rename(columns = {0:'khnp_buoy.wd'})

# ### df_kma_lightbecon_test_wd
# df_kma_lightbecon_test_wd = df_anch_drag_test.groupby(['year', 'month', 'day', 'hour', 'min'])['kma_lightbecon_test.wd'].agg(lambda x: pd.Series.mode(x)[0]).to_frame()
# df_kma_lightbecon_test_wd = df_kma_lightbecon_test_wd.T.unstack().reset_index()
# df_kma_lightbecon_test_wd.drop(columns = ['level_5'], inplace = True)
# df_kma_lightbecon_test_wd = df_kma_lightbecon_test_wd.rename(columns = {0:'kma_lightbecon.wd'})

# ### df_kma_lightbecon_test_ws
# df_kma_lightbecon_test_ws = round(df_anch_drag_test.groupby(['year', 'month', 'day', 'hour','min']).mean(),2)['kma_lightbecon_test.ws'].to_frame()
# df_kma_lightbecon_test_ws = df_kma_lightbecon_test_ws.T.unstack().reset_index()
# df_kma_lightbecon_test_ws.drop(columns = ['level_5'], inplace = True)
# df_kma_lightbecon_test_ws = df_kma_lightbecon_test_ws.rename(columns = {0:'kma_lightbecon.ws'})

# ### df_kma_pagobuoy_test_max_wh
# df_kma_pagobuoy_test_max_wh = round(df_anch_drag_test.groupby(['year', 'month', 'day', 'hour', 'min']).mean(),2)['kma_pagobuoy_test.max_wh'].to_frame()
# df_kma_pagobuoy_test_max_wh = df_kma_pagobuoy_test_max_wh.T.unstack().reset_index()
# df_kma_pagobuoy_test_max_wh.drop(columns = ['level_5'], inplace = True)
# df_kma_pagobuoy_test_max_wh = df_kma_pagobuoy_test_max_wh.rename(columns = {0:'kma_pagobuoy.max_wh'})

# ### df_kma_pagobuoy_test_sig_wh
# df_kma_pagobuoy_test_sig_wh = round(df_anch_drag_test.groupby(['year', 'month', 'day', 'hour', 'min']).mean(),2)['kma_pagobuoy_test.sig_wh'].to_frame()
# df_kma_pagobuoy_test_sig_wh = df_kma_pagobuoy_test_sig_wh.T.unstack().reset_index()
# df_kma_pagobuoy_test_sig_wh.drop(columns = ['level_5'], inplace = True)
# df_kma_pagobuoy_test_sig_wh = df_kma_pagobuoy_test_sig_wh.rename(columns = {0:'kma_pagobuoy.sig_wh'})

# ### df_kma_pagobuoy_test_mean_wh
# df_kma_pagobuoy_test_mean_wh = round(df_anch_drag_test.groupby(['year', 'month', 'day', 'hour', 'min']).mean(),2)['kma_pagobuoy_test.mean_wh'].to_frame()
# df_kma_pagobuoy_test_mean_wh = df_kma_pagobuoy_test_mean_wh.T.unstack().reset_index()
# df_kma_pagobuoy_test_mean_wh.drop(columns = ['level_5'], inplace = True)
# df_kma_pagobuoy_test_mean_wh = df_kma_pagobuoy_test_mean_wh.rename(columns = {0:'kma_pagobuoy.mean_wh'})

# ## Merge
# '''
# df_lat
# df_lon
# df_sog
# df_cog
# df_hdg
# df_khnp_buoy_test_ws
# df_khnp_buoy_test_wd
# df_kma_lightbecon_test_ws
# df_kma_lightbecon_test_wd
# df_kma_pagobuoy_test_max_wh
# df_kma_pagobuoy_test_sig_wh
# df_kma_pagobuoy_test_mean_wh
# '''
# df = pd.merge(df_lat, df_lon, how='inner')
# df = pd.merge(df, df_sog, how='inner')
# df = pd.merge(df, df_sog, how='inner')
# df = pd.merge(df, df_cog, how='inner')
# df = pd.merge(df, df_hdg, how='inner')
# df = pd.merge(df, df_khnp_buoy_test_ws, how='inner')
# df = pd.merge(df, df_khnp_buoy_test_wd, how='inner')
# df = pd.merge(df, df_kma_lightbecon_test_ws, how='inner')
# df = pd.merge(df, df_kma_lightbecon_test_wd, how='inner')
# df = pd.merge(df, df_kma_pagobuoy_test_max_wh, how='inner')
# df = pd.merge(df, df_kma_pagobuoy_test_sig_wh, how='inner')
# df = pd.merge(df, df_kma_pagobuoy_test_mean_wh, how='inner')

# # csv 저장
# df.to_csv("./통합데이터/df_test_v3.csv", index=False)

# # df
