==> 코드 정리/6. df_test_v4_mean.ipynb <==
import numpy as np
import pandas as pd

# df_test_v4_mean
- 테스트 데이터 중 정답 제출 양식에 맞는 데이터만 0인지, 1인지 모델 돌려서 predict하기 위한 데이터

df_test_v3_mean = pd.read_csv("./통합데이터/df_test_v3_mean.csv")
df_answer = pd.read_excel("./제출 양식/oceanUser.xlsx")

df_ulsan_answer = df_answer[(df_answer['AREA'] =='ULSAN')]

df = pd.merge(df_ulsan_answer, df_test_v3_mean, left_on = ['YEAR',	'MON',	'DAY',	'HOUR',	'MIN'], right_on = ['year', 'month', 'day', 'hour', 'min'], how ='left')
df.drop(columns=['AREA', 'NUM', 'YEAR', 'MON', 'DAY', 'HOUR', 'MIN', 'TARGET'], inplace=True)

# df.to_csv("../코드 정리/통합데이터/df_test_v4_mean.csv", index=False)

df
